+++ CHANGELOG.md
@@ -5,6 +5,11 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
+
+## **WORK IN PROGRESS**
+
+* (theimo1221) Move AutomaticFallbackTemperature to heatGroupSettings
+
 ## 2.17.4 (2023-10-18)
 
 * (theimo1221) Add includeInAmbientLight in fromPartialObject back-channel