+++ CHANGELOG.md
@@ -5,6 +5,11 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
+
+## **WORK IN PROGRESS**
+
+* (theimo1221) Minor tweaks to heating temperature handling
+
 ## 2.18.0 (2023-10-18)
 
 * (theimo1221) Move AutomaticFallbackTemperature to heatGroupSettings