+++ CHANGELOG.md
@@ -5,8 +5,7 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
-
-## **WORK IN PROGRESS**
+## 2.18.1 (2023-10-20)
 
 * (theimo1221) Minor tweaks to heating temperature handling